refactor(web): share series-scope toast digit mapping
Keep 1/2 ownership in one helper so quick-time yield and the toast
host cannot drift apart.
