+++ packages/web/src/events/recurrence/RecurrenceScopeOpportunityHost.tsx
@@ -11,27 +11,9 @@ import {
   selectRecurrenceScopeOpportunity,
   useRecurrenceScopeOpportunityStore,
 } from "@web/events/recurrence/recurrence-scope-opportunity.store";
+import { recurrenceScopeForToastDigit } from "@web/events/recurrence/recurrence-scope-toast-digit";
 import { isAppLocked } from "@web/shortcuts/app-lock";
-import { digitPickIndex } from "@web/shortcuts/digit-pick.util";
 import { isEditSequenceArmed } from "@web/shortcuts/useEditSequenceShortcut";
-
-const scopeForToastDigit = (
-  event: KeyboardEvent,
-): "thisAndFollowing" | "all" | null => {
-  if (
-    event.isComposing ||
-    isAppLocked() ||
-    isEditableKeyboardTarget(event) ||
-    isEditSequenceArmed()
-  ) {
-    return null;
-  }
-
-  const pickIndex = digitPickIndex(event);
-  if (pickIndex === 0) return "thisAndFollowing";
-  if (pickIndex === 1) return "all";
-  return null;
-};
 
 export function RecurrenceScopeOpportunityHost() {
   const opportunity = useRecurrenceScopeOpportunityStore(
@@ -43,7 +25,16 @@ export function RecurrenceScopeOpportunityHost() {
     const onKeyDown = (event: KeyboardEvent) => {
       if (event.defaultPrevented) return;
       if (!isRecurrenceScopeAskReady()) return;
-      const scope = scopeForToastDigit(event);
+      if (
+        event.isComposing ||
+        isAppLocked() ||
+        isEditableKeyboardTarget(event) ||
+        isEditSequenceArmed()
+      ) {
+        return;
+      }
+
+      const scope = recurrenceScopeForToastDigit(event);
       if (!scope) return;
 
       const current = useRecurrenceScopeOpportunityStore.getState().opportunity;
